修复抢券问题

+++ edj-market/src/main/java/com/edj/market/domain/vo/ActivityInfoVO.java
@@ -85,5 +85,5 @@ public class ActivityInfoVO {
      * 优惠券剩余数量
      */
     @Schema(description = "优惠券剩余数量")
-    private Long remainNum;
+    private Integer remainNum;
 }
